Vendoering revise
